% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeywordFinder.R
\name{KeywordFinder}
\alias{KeywordFinder}
\title{R6 class for keyword analysis}
\description{
TBD
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{KeywordFinder$new()}}
\item \href{#method-find_keywords}{\code{KeywordFinder$find_keywords()}}
\item \href{#method-score_sentiment}{\code{KeywordFinder$score_sentiment()}}
\item \href{#method-run}{\code{KeywordFinder$run()}}
\item \href{#method-print}{\code{KeywordFinder$print()}}
\item \href{#method-clone}{\code{KeywordFinder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeywordFinder$new(positive_keywords = NA, negative_keywords = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_keywords"></a>}}
\if{latex}{\out{\hypertarget{method-find_keywords}{}}}
\subsection{Method \code{find_keywords()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeywordFinder$find_keywords()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-score_sentiment"></a>}}
\if{latex}{\out{\hypertarget{method-score_sentiment}{}}}
\subsection{Method \code{score_sentiment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeywordFinder$score_sentiment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeywordFinder$run(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeywordFinder$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeywordFinder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
