% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentences.R
\name{sentence_segment}
\alias{sentence_segment}
\title{Segment sentences}
\usage{
sentence_segment(text)
}
\arguments{
\item{text}{String of clinical text.}
}
\description{
Convert text into a dataframe of sentences.
}
